+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/completion/providers/PrimaryExpr.kt
@@ -21,18 +21,28 @@ import com.intellij.psi.PsiElement
 import com.intellij.util.ProcessingContext
 import uk.co.reecedunn.intellij.plugin.core.completion.CompletionProviderEx
 import uk.co.reecedunn.intellij.plugin.intellij.resources.XQueryIcons
+import uk.co.reecedunn.intellij.plugin.xpath.completion.XPathEmptyFunctionInsertHandler
 import uk.co.reecedunn.intellij.plugin.xpath.completion.property.XPathCompletionProperty
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.EQNameCompletionType
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.completionType
 import uk.co.reecedunn.intellij.plugin.xpath.model.XPathVariableBinding
 import uk.co.reecedunn.intellij.plugin.xpath.model.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xquery.model.expand
+import uk.co.reecedunn.intellij.plugin.xquery.model.fileProlog
 import uk.co.reecedunn.intellij.plugin.xquery.model.inScopeVariables
+import uk.co.reecedunn.intellij.plugin.xquery.model.staticallyKnownFunctions
 
 fun createVariableLookup(localName: String, prefix: String?, element: PsiElement?): LookupElementBuilder {
     return LookupElementBuilder.create(prefix?.let { "$it:$localName" } ?: localName)
         .withIcon(XQueryIcons.Nodes.VarDecl)
         .withPsiElement(element)
+}
+
+fun createFunctionLookup(localName: String, prefix: String?, element: PsiElement?): LookupElementBuilder {
+    return LookupElementBuilder.create(prefix?.let { "$it:$localName" } ?: localName)
+        .withIcon(XQueryIcons.Nodes.FunctionDecl)
+        .withPsiElement(element)
+        .withInsertHandler(XPathEmptyFunctionInsertHandler)
 }
 
 object XQueryVarRefProvider : CompletionProviderEx {
@@ -89,6 +99,60 @@ object XQueryVarRefProvider : CompletionProviderEx {
                         val expanded = variable.variableName?.expand()?.firstOrNull()
                         if (expanded?.namespace?.data == varRef.namespace?.data) {
                             result.addElement(createVariableLookup(localName, null, variable.variableName?.element))
+                        }
+                    }
+                }
+            }
+            else -> {}
+        }
+    }
+}
+
+object XQueryFunctionCallProvider : CompletionProviderEx {
+    override fun apply(element: PsiElement, context: ProcessingContext, result: CompletionResultSet) {
+        val namespaces = context[XPathCompletionProperty.STATICALLY_KNOWN_FUNCTION_NAMESPACES]
+
+        val ref = element.parent as XsQNameValue
+        when (ref.completionType(element)) {
+            EQNameCompletionType.QNamePrefix, EQNameCompletionType.NCName -> {
+                // Without prefix context, so include all functions.
+                element.fileProlog()?.staticallyKnownFunctions()?.forEach { function ->
+                    val localName = function?.functionName?.localName?.data ?: return@forEach
+                    function.functionName?.expand()?.firstOrNull()?.let { name ->
+                        namespaces.forEach { ns ->
+                            if (ns.namespaceUri?.data == name.namespace?.data) {
+                                val declPrefix = ns.namespacePrefix?.data
+                                result.addElement(createFunctionLookup(localName, declPrefix, function.functionName?.element))
+                            }
+                        }
+                    }
+                }
+            }
+            EQNameCompletionType.QNameLocalName -> {
+                // With prefix context, so only include functions with a matching expanded QName namespace URI.
+                element.fileProlog()?.staticallyKnownFunctions()?.forEach { function ->
+                    val localName = function?.functionName?.localName?.data ?: return@forEach
+                    if (function.functionName?.prefix != null || function.functionName?.namespace != null) {
+                        function.functionName?.expand()?.firstOrNull()?.let { name ->
+                            namespaces.forEach { ns ->
+                                if (ns.namespaceUri?.data == name.namespace?.data) {
+                                    if (ns.namespacePrefix?.data == ref.prefix?.data) { // Prefix matches, and is already specified.
+                                        result.addElement(createFunctionLookup(localName, null, function.functionName?.element))
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            EQNameCompletionType.URIQualifiedNameLocalName -> {
+                // With namespace context, so only include functions with a matching expanded QName namespace URI.
+                element.fileProlog()?.staticallyKnownFunctions()?.forEach { function ->
+                    val localName = function?.functionName?.localName?.data ?: return@forEach
+                    if (function.functionName?.prefix != null || function.functionName?.namespace != null) {
+                        val expanded = function.functionName?.expand()?.firstOrNull()
+                        if (expanded?.namespace?.data == ref.namespace?.data) {
+                            result.addElement(createFunctionLookup(localName, null, function.functionName?.element))
                         }
                     }
                 }